+++ services/explorer-ui-v2/src/pages/l1-events/index.tsx
@@ -4,6 +4,12 @@ import { CopyableAddress, Pagination } from "~/components/common";
 import { ConsoleHead, Shell } from "~/components/layout";
 import { useChainInfo, useL1ContractEvents } from "~/hooks/api";
 import { ageStr, fmtNum, truncateHashString } from "~/lib/utils";
+import {
+  argsPreview,
+  eventKind,
+  eventTimestampMs,
+  findL2BlockNumber,
+} from "./helpers";
 
 const PAGE_SIZE = 25;
 
@@ -54,69 +60,6 @@ const CONTRACT_TABS: ContractTab[] = [
     desc: "Canonical registry of rollup versions & addresses.",
   },
 ];
-
-/**
- * Colour + semantic bucket for an event name. Events that don't match a known
- * pattern fall into the neutral bucket with a grey bullet.
- */
-const EVENT_KIND = (eventName: string): string => {
-  const n = eventName.toLowerCase();
-  if (n.includes("block") || n.includes("proof") || n.includes("epoch"))
-    {return "block";}
-  if (n.includes("message") || n.includes("inbox") || n.includes("outbox") || n.includes("root"))
-    {return "message";}
-  if (n.includes("deposit")) {return "deposit";}
-  if (n.includes("withdraw")) {return "withdraw";}
-  if (n.includes("slash")) {return "slash";}
-  if (n.includes("validator") || n.includes("attester")) {return "validator";}
-  if (n.includes("rollup") || n.includes("registry")) {return "registry";}
-  return "other";
-};
-
-/** Human-ish preview of eventArgs — first 2 keys, values truncated. */
-const argsPreview = (
-  args: ChicmozL1GenericContractEvent["eventArgs"],
-): string => {
-  if (!args) {return "—";}
-  const entries = Object.entries(args).slice(0, 2);
-  if (entries.length === 0) {return "—";}
-  return entries
-    .map(([k, v]) => `${k}=${formatArgValue(v)}`)
-    .join(" · ");
-};
-
-const formatArgValue = (v: unknown): string => {
-  if (v === null || v === undefined) {return "null";}
-  if (typeof v === "bigint") {return v.toString();}
-  if (typeof v === "string") {
-    if (v.startsWith("0x") && v.length > 20)
-      {return truncateHashString(v, 8, 6);}
-    return v;
-  }
-  if (typeof v === "boolean" || typeof v === "number") {return String(v);}
-  return "…";
-};
-
-const findL2BlockNumber = (
-  args: ChicmozL1GenericContractEvent["eventArgs"],
-): number | null => {
-  if (!args) {return null;}
-  const candidates = ["l2BlockNumber", "l2Block", "blockNumber"];
-  for (const k of candidates) {
-    const v = (args as Record<string, unknown>)[k];
-    if (typeof v === "bigint") {return Number(v);}
-    if (typeof v === "number" && Number.isFinite(v)) {return v;}
-    if (typeof v === "string" && /^\d+$/.test(v)) {return Number(v);}
-  }
-  return null;
-};
-
-const eventTimestampMs = (
-  evt: ChicmozL1GenericContractEvent,
-): number | null => {
-  const ts = evt.l1BlockTimestamp;
-  return typeof ts === "number" && Number.isFinite(ts) ? ts : null;
-};
 
 export const L1EventsPage: FC = () => {
   const { data: chainInfo } = useChainInfo();
@@ -455,7 +398,7 @@ export const L1EventsPage: FC = () => {
         <div>
           {paged.map((e) => {
             const l2Block = findL2BlockNumber(e.eventArgs);
-            const kind = EVENT_KIND(e.eventName);
+            const kind = eventKind(e.eventName);
             const ts = eventTimestampMs(e);
             const txHref = e.l1TransactionHash
               ? `https://etherscan.io/tx/${e.l1TransactionHash}#eventlog`
